refactor: move keyboard hook to common
