+++ apps/dashboard/app/(app)/logs-v2/components/controls/components/logs-display/components/display-popover.tsx
@@ -1,7 +1,7 @@
 import { isDisplayProperty, useLogsContext } from "@/app/(app)/logs-v2/context/logs";
-import { useKeyboardShortcut } from "@/app/(app)/logs-v2/hooks/use-keyboard-shortcut";
 import { KeyboardButton } from "@/components/keyboard-button";
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
+import { useKeyboardShortcut } from "@/hooks/use-keyboard-shortcut";
 import {
   type KeyboardEvent,
   type PropsWithChildren,
